+++ plaso/filters/expressions.py
@@ -6,9 +6,11 @@ from __future__ import unicode_literals
 import abc
 import logging
 
+from dfdatetime import posix_time as dfdatetime_posix_time
+
 from plaso.lib import errors
+from plaso.lib import py2to3
 from plaso.filters import filters
-from plaso.filters import helpers
 
 
 class Expression(object):
@@ -174,6 +176,41 @@ class EventExpression(Expression):
     super(EventExpression, self).__init__()
     self._bool_value = True
 
+  def _CopyValueToDateTime(self, value):
+    """Copies an event filter value to a date and time object.
+
+    Args:
+      value (str): event filter value.
+
+    Returns:
+      dfdatetime.PosixTimeInMicroseconds: date and time object.
+
+    Raises:
+      ValueError: if the value cannot be copied to a date and time object.
+    """
+    if not isinstance(value, py2to3.INTEGER_TYPES):
+      try:
+        value = int(value, 10)
+      except (TypeError, ValueError):
+        pass
+
+    if isinstance(value, py2to3.INTEGER_TYPES):
+      date_time = dfdatetime_posix_time.PosixTimeInMicroseconds(
+          timestamp=value)
+    else:
+      try:
+        # Adjust the ISO 8601 string so is rembles a Python date and time
+        # string.
+        if value and len(value) > 10 and value[10] == 'T':
+          value = ' '.join(value.split('T'))
+
+        date_time = dfdatetime_posix_time.PosixTimeInMicroseconds()
+        date_time.CopyFromDateTimeString(value)
+      except (TypeError, ValueError):
+        raise ValueError('Unsupported timestamp value: {0!s}'.format(value))
+
+    return date_time
+
   def Compile(self):
     """Compiles the expression into a filter.
 
@@ -204,7 +241,7 @@ class EventExpression(Expression):
       date_time_arguments = []
       for argument in self.args:
         try:
-          date_time = helpers.CopyValueToDateTime(argument)
+          date_time = self._CopyValueToDateTime(argument)
         except ValueError as exception:
           raise errors.ParseError(exception)
